% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.ab.R
\name{beta.ab}
\alias{beta.ab}
\title{Get Beta parameters}
\usage{
beta.ab(m, v)
}
\arguments{
\item{m}{mean efficacy of a dose (single value). Values range from 0 - 100. (e.g, T cell persistence - values b/w 5 and 80 per cent)}

\item{v}{efficacy variances of a dose (single value). Values range from 0 - 1. (e.g., 0.01)}
}
\description{
Function beta.ab() returns parameters alpha and beta for generating beta r.v. (per dose)
}
