% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.stg1.R
\name{gen.stg1}
\alias{gen.stg1}
\title{Generate stage 1 - safety}
\usage{
gen.stg1(dose, dose.tox, p0, p1, K, coh.size, m, v, nbb = 100)
}
\arguments{
\item{dose}{number of doses to be tested (scalar)}

\item{dose.tox}{vector of true toxicities for each dose. Values range from 0 - 1.}

\item{p0}{toxicity under null (unsafe DLT rate). Values range from 0 - 1.}

\item{p1}{toxicity under alternative (safe DLT rate). Values range from 0 - 1; p0 > p1}

\item{K}{threshold for LR. Takes integer values: 1,2,... (recommended K=2)}

\item{coh.size}{cohort size (number of patients) per dose}

\item{m}{mean efficacy of a dose (single value). Values range from 0 - 100. (e.g, T cell persistence - values b/w 5 and 80 per cent)}

\item{v}{efficacy variances of a dose (single value). Values range from 0 - 1. (e.g., 0.01)}

\item{nbb}{binomial parameter (default = 100 cells per patient)}
}
\value{
List of the following objects:
           Y.alloc - vector of Ys corresponding to all dose assign. from stage 1 
                     (acceptable and unacceptable)
           d.alloc - vector of all dose assign. from stage 1
            Y.safe - subset of Y.alloc corresponding to only acceptable doses; to be used for 
                     randomization in stage 2
            d.safe - subset of d.alloc corresponding to only acceptable dose assign.; to be used for 
                     randomization in stage 2
         tox. safe - vector of toxicities corresponding to acceptable doses
                n1 - total number of pts. allocated in stage 1
}
\description{
Function gen.y.a() uses a beta-binomial distribution to generate outcomes (Ys) corresponding 
 to all dose assignment in stage 1.
}
\examples{
write examples here

}
