% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.stg2.R
\name{gen.stg2}
\alias{gen.stg2}
\title{Simulate stage 2 - efficacy}
\usage{
gen.stg2(dose, dose.tox, p0, p1, K, coh.size, m, v, N, stop.rule = 9,
  cohort = 1, samedose = T, nbb = 100)
}
\arguments{
\item{dose}{- number of doses to be tested (scalar)}

\item{dose.tox}{- vector of true toxicities for each dose. Values range from 0 - 1.}

\item{p0}{- toxicity under null (unsafe DLT rate). Values range from 0 - 1.}

\item{p1}{- toxicity under alternative (safe DLT rate). Values range from 0 - 1; p0 > p1}

\item{K}{- threshold for LR. Takes integer values: 1,2,... (recommended K=2)}

\item{coh.size}{- cohort size (number of patients) per dose}

\item{m}{- mean efficacy of a dose (single value). Values range from 0 - 100. (e.g, T cell persistence - values b/w 5 and 80 per cent)}

\item{v}{- efficacy variances of a dose (single value). Values range from 0 - 1. (e.g., 0.01)}

\item{N}{- max sample size for stages 1&2}

\item{stop.rule}{- if only dose 1 safe, allocate up to 9 (default) patients at dose 1 to collect more info}

\item{nbb}{- binomial parameter (default = 100 cells per patient)}

\item{nbb}{- binomial parameter (default = 100 cells per patient)}
}
\description{
Function f.study.a() fits a linear regression for the continuous efficacy outcomes,
computes the randomization probabilities/dose and allocates the next patient to a dose that
is considered acceptably safe and has the highest efficacy
Returns: yk.final - vector of all efficacy outcomes (Ys) corresponding to dose assign.
                     (Stage 1&2)
         dk.final - vector of all dose assign.(Stage 1&2)
If no dose allocation, put NAs in dk.final and yk.final
}
