% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox.stop.R
\name{tox.stop}
\alias{tox.stop}
\title{Generate DLTs and calculate the LR for each dose}
\usage{
tox.stop(dose, dose.tox, p0, p1, K, coh.size)
}
\arguments{
\item{dose}{- number of doses to be tested (scalar)}

\item{dose.tox}{- vector of true toxicities for each dose. Values range from 0 - 1.}

\item{p0}{- toxicity under null (unsafe DLT rate). Values range from 0 - 1.}

\item{p1}{- toxicity under alternative (safe DLT rate). Values range from 0 - 1; p0 > p1}

\item{K}{- threshold for LR. Takes integer values: 1,2,...(recommended K=2)}
}
\description{
Returns a 4-column matrix containing dose assign., DLTs at each dose, cohort number, and LR
}
