% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.ab.R
\name{beta.ab}
\alias{beta.ab}
\title{Generates parameters for the beta distribution # I don't think we need to show this as a separate function, 
                                                        but put together with gen.eff.stg1 or be called by gen.eff.stg1}
\usage{
beta.ab(m, v)
}
\arguments{
\item{m}{Vector of mean efficacies per dose. Values range from 0 - 100. (e.g, T cell persistence - values b/w 5 and 80 per cent)}

\item{v}{Vector of efficacy variances per dose. Values range from 0 - 1. (e.g., 0.01)}
}
\value{
Vector of alpha and beta values for generating beta random variable for a dose.
}
\description{
Function \code{beta.ab()} returns parameters alpha and beta for generating beta r.v. (per dose)
}
\keyword{internal}
