% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eff.stg1.R
\name{eff.stg1}
\alias{eff.stg1}
\title{Generates efficacy outcomes for stage 1}
\usage{
eff.stg1(dose, dose.tox, p1, p2, K, coh.size, m, v, nbb = 100)
}
\arguments{
\item{dose}{number of doses to be tested (scalar)}

\item{dose.tox}{vector of true toxicities for each dose. Values range from 0 - 1.}

\item{p1}{toxicity under null (unsafe DLT rate). Values range from 0 - 1.}

\item{p2}{toxicity under alternative (safe DLT rate). Values range from 0 - 1; p1 > p2}

\item{K}{threshold for LR. Takes integer values: 1,2,...(recommended K=2)}

\item{coh.size}{cohort size (number of patients) per dose (Stage 1)}

\item{m}{vector of mean efficacies per dose. Values range from 0 - 100. (e.g, T cell persistence - values b/w 5 and 80 per cent)}

\item{v}{vector of efficacy variances per dose. Values range from 0 - 1. (e.g., 0.01)}

\item{nbb}{binomial parameter (default = 100 cells per patient)}
}
\description{
Function \code{eff.stg1()} uses a beta-binomial distribution to generate outcomes (Ys) corresponding 
to acceptable dose assignments from stage 1
}
\examples{

}
