% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe.dose.R
\name{safe.dose}
\alias{safe.dose}
\title{Identify safe doses}
\usage{
safe.dose(dose, dose.tox, p0, p1, K, coh.size)
}
\arguments{
\item{dose}{- number of doses to be tested (scalar)}

\item{dose.tox}{- vector of true toxicities for each dose. Values range from 0 - 1.}

\item{p0}{- toxicity under null (unsafe DLT rate). Values range from 0 - 1.}

\item{p1}{- toxicity under alternative (safe DLT rate). Values range from 0 - 1; p0 > p1}

\item{K}{- threshold for LR. Takes integer values: 1,2,...(recommended K=2)}

\item{coh.size}{- cohort size (number of patients) per dose}
}
\description{
Function safe.dose() distinguishes acceptable from unacceptable doses
Returns: alloc.safe - matrix of assign. for only acceptable doses (to be used in stage 2) and the 
                      corresponding toxicities
         alloc.total - vector of all assign. from stage 1 
         n1 - total number of pts. allocated in stage 1
}
