% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox.profile.R
\name{tox.profile}
\alias{tox.profile}
\title{Generates DLTs and calculate the likelihood-ratio (LR) for each dose}
\usage{
tox.profile(dose, dose.tox, p1, p2, K, coh.size)
}
\arguments{
\item{dose}{number of doses to be tested (scalar)}

\item{dose.tox}{vector of true toxicities for each dose. Values range from 0 - 1.}

\item{p1}{toxicity under null (unsafe DLT rate). Values range from 0 - 1.}

\item{p2}{toxicity under alternative (safe DLT rate). Values range from 0 - 1; p1 > p2}

\item{K}{threshold for LR. Takes integer values: 1,2,...(recommended K=2)}

\item{coh.size}{cohort size (number of patients) per dose (Stage 1)}
}
\value{
4-column matrix containing dose assign., DLTs at each dose, cohort number, and LR # This is the same is descripton above. Keep only one?
}
\description{
Returns a 4-column matrix containing dose assignment, DLTs at each dose, cohort number, and LR
}
\examples{
dose = 5                                      # Dose levels
dose.tox <- c(0.05, 0.10, 0.15, 0.20, 0.30)   # True toxicity per dose
p1 = 0.40                                     # Unacceptable DLT rate
p2 = 0.15                                     # Acceptable DLT rate
K = 2                                         # Likelihood-ratio (LR) threshold
coh.size = 3                                  # Assign 3 pts per dose in stage 1
}
