% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tox.stop.R
\name{tox.stop}
\alias{tox.stop}
\title{Generate DLTs and calculate the LR for each dose}
\usage{
tox.stop(dose, dose.tox, p0, p1, K, coh.size)
}
\arguments{
\item{dose}{- number of doses to be tested (scalar)}

\item{dose.tox}{- vector of true toxicities for each dose. Values range from 0 - 1.}

\item{p0}{toxicity under null (unsafe DLT rate). Values range from 0 - 1.}

\item{p1}{toxicity under alternative (safe DLT rate). Values range from 0 - 1; p0 > p1}

\item{K}{threshold for LR. Takes integer values: 1,2,...(recommended K=2)}

\item{coh.size}{cohort size (number of patients) per dosetox.stop <- function(dose, dose.tox, p0, p1, K, coh.size)}
}
\value{
4-column matrix containing dose assign., DLTs at each dose, cohort number, and LR
}
\description{
Returns a 4-column matrix containing dose assign., DLTs at each dose, cohort number, and LR
}
\examples{
dose = 5                                      # dose levels
dose.tox <- c(0.05, 0.10, 0.15, 0.20, 0.30)   # True toxicity per dose
p0 = 0.40                                     # Unacceptable DLT rate
p1 = 0.15                                     # Acceptable DLT rate
K = 2                                         # Likelihood-ratio (LR) threshold
coh.size = 3                                  # (3 pts per dose in stage 1)
m <- c(5, 15, 40, 65, 80)                     # Mean efficacy per dose (here mean persistence per dose (per cents)
v <- rep(0.01,5)                              # Efficacy variance per dose
N = 25                                        # Total sample size (stages 1&2)
stop.rule = 9                                 # If dose 1 is the only safe dose remaining, allocate up to 9 pts. to collect more info

tox.stop(dose = dose, dose.tox = dose.tox, p0 = p0, p1 = p1, K = K, coh.size = coh.size)

}
